🚚 refactor: move the feature folder
