+++ src/config/featureFlags/utils/parser.ts
@@ -1,4 +1,4 @@
-import { FeatureFlagsSchema, IFeatureFlags } from '@/config/featureFlags';
+import { FeatureFlagsSchema, IFeatureFlags } from '../schema';
 
 /**
  * 解析环境变量中的特性标志字符串。
